Updated README and uploaded PDF. Closes #4 and #5
