+++ README.md
@@ -9,4 +9,4 @@ Programming.
 ## Materials
 - [VS project with .cs files](oopconcepts)
 - [UML diagrams](materials/UML-Diagrams)
-- Presentation in both [online](https://www.canva.com/design/DAGYFe4lUmY/x9nbx1vKyC_-RK8nQx_hiQ/view?utm_content=DAGYFe4lUmY) and [PDF]() formats.
+- Presentation in both [online](https://www.canva.com/design/DAGYFe4lUmY/x9nbx1vKyC_-RK8nQx_hiQ/view?utm_content=DAGYFe4lUmY) and [PDF](materials/OOP%20Concepts%20Presentation%20-%20Alvaro%20Cobo.pdf) formats.
