--- oopconcepts/Building.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace oopconcepts
-{
-    public class Building
-    {
-        public List<Door> Doors { get; set; } = new List<Door>();
-        //We haven't seen this yet, but it's a list of doors
-    }
-}
