--- oopconcepts/Key.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace oopconcepts
-{
-    public class Key
-    {
-        public int KeyID { get; set; }
-
-        public Key(int keyID)
-        {
-            KeyID = keyID;
-        }
-    }
-}
