--- oopconcepts/Wall.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace oopconcepts
-{
-    public class Wall
-    {
-        public string Color { get; set; }
-    }
-}
