+++ oopconcepts/classes/Key.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace oopconcepts
+{
+    public class Key
+    {
+        public int KeyID { get; set; }
+
+        public Key(int keyID)
+        {
+            KeyID = keyID;
+        }
+    }
+}
