+++ oopconcepts/classes/Room.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace oopconcepts
+{
+    public class Room
+    {
+        public Wall[] Walls { get; set; }
+
+        public Room(Wall[] walls)
+        {
+            Walls = walls;
+        }
+    }
+}
