+++ oopconcepts/classes/Wall.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace oopconcepts
+{
+    public class Wall
+    {
+        public string Color { get; set; }
+    }
+}
